Fixed apache conf, copying default settings only if needed

+++ scripts/create_db.sh
@@ -30,8 +30,12 @@ else
 
 fi
 
+
 mkdir -p settings
-cp signetsim/settings/settings.json settings/settings.json
+if [ ! -f ${INSTALL_DIR}/settings/settings.json ]
+then
+    cp signetsim/settings/settings.json settings/settings.json
+fi
 
 chgrp -R www-data data
 chmod -R 664 data
